Logs:
-> Configured @POST endpoint to send the pokemon payload to the API for capturing.

-> For each capture, the local team and captured tables are updated with a new entry.

+++ app/src/main/java/com/sivan/pokebolt/data/CapturedItem.kt
@@ -1,8 +1,6 @@
 package com.sivan.pokebolt.data
 
 import android.os.Parcelable
-import androidx.room.ColumnInfo
-import androidx.room.PrimaryKey
 import com.sivan.pokebolt.retrofit.entity.Moves
 import com.sivan.pokebolt.retrofit.entity.Sprites
 import com.sivan.pokebolt.retrofit.entity.Stats
@@ -28,8 +26,8 @@ data class CapturedItem(
 
     var moves: List<Moves>,
 
-    var captured_lat_at: Double,
+    var captured_lat_at: Float,
 
-    var captured_long_at: Double,
+    var captured_long_at: Float,
 ) : Parcelable{
 }
