+++ app/src/main/java/com/sivan/pokebolt/retrofit/entity/CapturedResponse.kt
@@ -27,10 +27,10 @@ data class CapturedResponse(
 
     @SerializedName("captured_lat_at")
     @Expose
-    val captured_lat_at: Double,
+    val captured_lat_at: Float,
 
     @SerializedName("captured_long_at")
     @Expose
-    val captured_long_at: Double
+    val captured_long_at: Float
 ) : Parcelable {
 }
